// packages/web/src/components/input/interface.ts
export interface TextAreaProps
  extends Omit<JSX.TextareaHTMLAttributes<HTMLTextAreaElement>, 'onInput' | 'maxLength'> {
  /**
   * @zh 开启字数统计之后，会在 `textarea` 标签外包一层 `div`，`wrapperStyle` 用来配置这个 `div` 的样式。
   * @en With `showWordLimit`, a `div` will be outside the `textarea` tag, and `wrapperStyle` is used to configure the style of it.
   */
  wrapperStyle?: JSX.CSSProperties
  /**
   * @zh 是否禁用
   * @en Whether the textarea is disabled
   */
  disabled?: boolean
  /**
   * @zh 默认值
   * @en To set default value
   */
  defaultValue?: string
  /**
   * @zh 值
   * @en To set value
   */
  value?: string
  /**
   * @zh 是否自动调整输入框的高度
   * @en Height autoSize feature
   */
  autoSize?: boolean | { minRows?: number; maxRows?: number }
  /**
   * @zh 是否是错误状态。(废弃，下个大版本移除，使用 status='error' 替代)
   * @en Whether the textarea is error.(Deprecated, removed in the next major version, use status='error' instead)
   * @deprecated
   */
  error?: boolean
  /**
   * @zh 状态
   * @en Status
   * @version 2.45.0
   */
  status?: 'error' | 'warning'
  /**
   * @zh 输入框提示文字
   * @en textarea placeholder
   */
  placeholder?: string
  /**
   * @zh 输入时的回调
   * @en Callback when user input
   */
  onInput?: (value: string, e: InputEvent) => void
  /**
   * @zh 按下回车键的回调
   * @en Callback when press enter key
   */
  onPressEnter?: (e: KeyboardEvent) => void
  /**
   * @zh 输入框最大输入的长度；设置 `errorOnly`为 `true` 后，超过 `maxLength` 会展示 `error` 状态，并不限制用户输入。
   * @en The max content length；After setting `errorOnly` to `true`, if `maxLength` is exceeded, the `error` status will be displayed, and user input will not be restricted.
   * @version `errorOnly` in 2.23.0
   */
  maxLength?: number | { length: number; errorOnly?: boolean }
  showWordLimit?: boolean
  /**
   * @zh 允许清空输入框
   * @en Whether allow clear the content
   * @version 2.2.0
   */
  allowClear?: boolean
  /**
   * @zh 点击清除按钮的回调
   * @en Callback when click clear button
   * @version 2.2.0
   */
  onClear?: () => void
  /**
   * @zh `allowClear` 时配置清除按钮的图标。
   * @en Configure the icon of the clear button when `allowClear`.
   * @version 2.50.0
   */
  clearIcon?: JSXElement
}